Оптимизация вытяга с дб

+++ ISPH.Infrastructure/Repositories/EmployersRepository.cs
@@ -24,10 +24,7 @@ namespace ISPH.Infrastructure.Repositories
         }
         public override async Task<Employer> GetById(int id)
         {
-            var employer = await _context.Employers.FindAsync(id);
-            if (employer == null) return null;
-            employer.Advertisements = await _context.Advertisements.AsQueryable().Where(adv => adv.EmployerId == id).ToListAsync();
-            return employer;
+            return await _context.Employers.AsNoTracking().Include(emp => emp.Advertisements).FirstOrDefaultAsync(adv => adv.EmployerId == id);
         }
         public override async Task<bool> HasEntity(Employer entity)
         {
@@ -45,13 +42,15 @@ namespace ISPH.Infrastructure.Repositories
 
         public async Task<bool> UpdateCompany(Employer entity, string companyName)
         {
-            var ads = _context.Advertisements.AsQueryable().Include(ad => ad.Employer).Where(ad => ad.Employer.EmployerId == entity.EmployerId);
             var company = await _context.Companies.FirstOrDefaultAsync(com => com.Name == companyName);
-            entity.CompanyName = companyName;
-            entity.CompanyId = company.CompanyId;
-            entity.Company = company;
-            _context.Employers.Update(entity);
-            _context.Advertisements.RemoveRange(ads);
+            if (company != null)
+            {
+                entity.CompanyName = companyName;
+                entity.CompanyId = company.CompanyId;
+                _context.Employers.Update(entity);
+                var ads = _context.Advertisements.AsQueryable().Where(ad => ad.EmployerId == entity.EmployerId);
+                _context.Advertisements.RemoveRange(ads);
+            }
             return await _context.SaveChangesAsync() > 0;
         }
         //Auth
@@ -61,8 +60,9 @@ namespace ISPH.Infrastructure.Repositories
             hashService.CreateHashedPassword(password, out byte[] hashedPass, out byte[] SaltPass);
             user.HashedPassword = hashedPass;
             user.SaltPassword = SaltPass;
-            if (await Create(user)) return await _context.Employers.FirstOrDefaultAsync(emp => emp.Email == user.Email);
-            return null;
+            _context.Employers.Add(user);
+            await _context.SaveChangesAsync();
+            return user;
         }
 
         public async Task<Employer> Login(string email, string password)
